refactored to use in kotlin multiplatform projects also upgraded gradle and project sructure to current standards
