--- build.gradle
@@ -1,15 +1,0 @@
-plugins {
-    id 'org.jetbrains.kotlin.jvm' version '1.3.50'
-}
-
-group 'mm'
-version '0.1'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib"
-}
-
